MIP and Aliveness Monitor Improvements
During integration testing we following issues were found:
(a) When OVS disconnects, Aliveness monitor tries to send ARP request and
fails. This leads to withdrawal of MIP IP from DCGW and data-path is
disrupted.
(b) Aliveness Monitor is not started post-upgrade or cluster-reboot

Fixes:
(a) When OVS disconnects/connects, Aliveness monitor will be
paused/unpaused by L3VPN. For this new container is defined to store the
mapping between DPN to List of Mips.

(b) When ODL reboots, to start the aliveness monitor again , MIP info has
to be persisted.LearntVpnVipToPort is Operational DS, which gets wiped
out after reboot.NeutronVpnVipToPort is a config DS which already stores
the MIP-IP. Hence LearntVpnVipToPort being a duplicate info , is removed.
New NeutronVpnVipToPort listener is implemented so that, after reboot ,
MIP is replayed and aliveness monitor is started automatically.

After reboot, when NeutronVpnVipToPort data is replayed , L3VPN starts
Monitor Task. But AlivenessManager fails to send ARP because VM port may
not be UP. Hence only when InterfaceState is ready for the VM port ,
Monitor Task is started. Also after ODL reboot , MIP-adj disappears from
config-vpn-intf due to update done by NeutronVpn on vpn-interface.Hence it
must be updated after reboot.

AlivenessManager needs NeutronVpnService to get GatewayMacAddr. But After
reboot ,
vpnmanager-impl bundle comes up first before NeutronVpnService(i.e
neutronvpn-impl) bundle. Due to this problem, AlivenessManager fails to
send Arp after ODL reboot. Hence to resolve this, MacEntry is enhanced to
provide gatewayIP to directly to remove the dependency.

+++ vpnmanager/impl/src/main/java/org/opendaylight/netvirt/vpnmanager/iplearn/IpMonitorStartTask.java
@@ -12,13 +12,14 @@ import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.Callable;
 import org.opendaylight.netvirt.vpnmanager.iplearn.model.MacEntry;
+import org.opendaylight.yangtools.yang.common.Uint32;
 
 public class IpMonitorStartTask implements Callable<List<ListenableFuture<Void>>> {
     private final MacEntry macEntry;
-    private final Long arpMonitorProfileId;
+    private final Uint32 arpMonitorProfileId;
     private final AlivenessMonitorUtils alivenessMonitorUtils;
 
-    public IpMonitorStartTask(MacEntry macEntry, Long profileId, AlivenessMonitorUtils alivenessMonitorUtils) {
+    public IpMonitorStartTask(MacEntry macEntry, Uint32 profileId, AlivenessMonitorUtils alivenessMonitorUtils) {
         this.macEntry = macEntry;
         this.arpMonitorProfileId = profileId;
         this.alivenessMonitorUtils = alivenessMonitorUtils;
